+++ CHANGELOG.md
@@ -12,7 +12,8 @@ date.
 ## System Administration Changes
 
 ### Added
-* Added `--mindreader-max-console-length-in-characters` which is the limit in characters that we allow action trace's console output to be before truncating them.
+
+* Added `--mindreader-max-console-length-in-bytes` which is the limit in bytes that we allow action trace's console output to be before truncating them.
 * Environment variable `MINDREADER_MAX_TOKEN_SIZE` can now be set to override `bufio.Scanner()` max token size (default `52428800`, i.e. `50Mb`) for EOSIO chains with huge transactions
 * Flag `--accounthist-mode` to specific the accounthist mode of operation
 * Added `tools check accounthist-shards` to